+++ src/array/chunk_grid.rs
@@ -53,6 +53,7 @@ impl ChunkGrid {
         }
         Err(PluginCreateError::Unsupported {
             name: metadata.name().to_string(),
+            plugin_type: "chunk grid".to_string(),
         })
     }
 }