+++ src/array/chunk_key_encoding.rs
@@ -44,6 +44,7 @@ impl ChunkKeyEncoding {
         }
         Err(PluginCreateError::Unsupported {
             name: metadata.name().to_string(),
+            plugin_type: "chunk key encoding".to_string(),
         })
     }
 }