+++ src/array/codec/bytes_to_bytes/gzip/gzip_codec.rs
@@ -11,7 +11,7 @@ use crate::{
         BytesRepresentation,
     },
     metadata::Metadata,
-    plugin::PluginCreateError,
+    plugin::{PluginCreateError, PluginMetadataInvalidError},
 };
 
 #[cfg(feature = "async")]
@@ -35,7 +35,9 @@ fn is_name_gzip(name: &str) -> bool {
 }
 
 fn create_codec_gzip(metadata: &Metadata) -> Result<Codec, PluginCreateError> {
-    let configuration: GzipCodecConfiguration = metadata.to_configuration()?;
+    let configuration: GzipCodecConfiguration = metadata
+        .to_configuration()
+        .map_err(|_| PluginMetadataInvalidError::new(IDENTIFIER, "codec", metadata.clone()))?;
     let codec = Box::new(GzipCodec::new_with_configuration(&configuration));
     Ok(Codec::BytesToBytes(codec))
 }